+++ client/src/components/forms/dynamic-form/components.js
@@ -319,4 +319,5 @@ export {
   DynamicMultiSelect,
   DynamicPublicKey,
   DynamicDatePicker,
+  tagColors,
 };